requirements.txt
streamlit
pandas
scikit-learn
numpy
PyPDF2
python-docx
uvicorn
fastapi